+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ### Changed
 _(nothing)_
 ### Added
-_(nothing)_
+- iron.closure-utils/debug? - Are we in dev or production release?
 ### Removed
 _(nothing)_
 ### Fixed